push actual code for error searching

+++ bc-truffle/src/main/java/node/BcTypes.java
@@ -2,7 +2,10 @@ package node;
 
 import com.oracle.truffle.api.dsl.TypeSystem;
 
-@TypeSystem({long.class, String.class})
+@TypeSystem({
+        double.class,
+        String.class
+})
 public abstract class BcTypes {
 
 }
